Revert "SERVER-43762 tighten the overload set for BSONObjBuilder::appendNumber"
This reverts commit 98f42065a3f530024980e72ef0a15a371c28e32b.

GitOrigin-RevId: 07a2f1b3245d2a18a8b53482091aa32cbf9210be

+++ src/mongo/db/modules/enterprise/src/search/mongotmock/mongotmock_ismaster.cpp
@@ -53,7 +53,7 @@ public:
 
         result.appendBool("ismongot", true);
         result.appendNumber("maxBsonObjectSize", BSONObjMaxUserSize);
-        result.appendNumber("maxMessageSizeBytes", static_cast<long long>(MaxMessageSizeBytes));
+        result.appendNumber("maxMessageSizeBytes", MaxMessageSizeBytes);
         result.appendDate("localTime", jsTime());
 
         auto wireSpec = WireSpec::instance().get();
